requirements.txt
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
djangorestframework==3.14.0
#xhtml2pdf==0.2.10
